+++ LICENSES
@@ -44,6 +44,7 @@ otherwise.
 - static/themes/default/img/network-wireless.svg :: [[http://tango.freedesktop.org/][Public Domain]]
 - static/themes/default/icons/deluge.png :: [[https://upload.wikimedia.org/wikipedia/commons/thumb/8/85//Deluge-Logo.svg/2000px-Deluge-Logo.svg.png][GPL]]
 - static/themes/default/icons/diaspora.png :: [[https://upload.wikimedia.org/wikipedia/commons/thumb/8/85//Deluge-Logo.svg/2000px-Deluge-Logo.svg.png][Publc Domain]]
+- static/themes/default/icons/ejabberd.png :: [[https://www.ejabberd.im/][GPL-2]]
 - static/themes/default/icons/infinoted.png :: [[https://github.com/gobby/gobby/blob/master/COPYING][ISC]]
 - static/themes/default/icons/ikiwiki.png :: [[https://ikiwiki.info/][GPL-2+]]
 - static/themes/default/icons/jsxc.png :: -
@@ -58,10 +59,6 @@ otherwise.
 - static/themes/default/icons/roundcube.png :: [[https://roundcube.net/][GPL-3+]]
 - static/themes/default/icons/shaarli.png :: [[https://github.com/shaarli/Shaarli][zlib/libpng]]
 - static/themes/default/icons/syncthing.png :: [[https://github.com/syncthing/syncthing/][Mozilla Public License Version 2.0]]
+- static/themes/default/icons/tahoe.png :: [[https://en.wikipedia.org/w/index.php?curid=31565848][Creative Commons Attribution 3.0 Unported License]]
 - static/themes/default/icons/transmission.png :: [[https://transmissionbt.com/][GPL]]
 - static/themes/default/icons/ttrss.png :: [[https://tt-rss.org/gitlab/fox/tt-rss][GPL]]
-- static/themes/default/icons/ejabberd.png :: [[https://www.ejabberd.im/][GPL-2]]
-- static/themes/default/icons/xmpp.png :: [[https://www.ejabberd.im/][GPL3]]
-- static/themes/default/icons/openvpn.png :: [[https://github.com/thekishanraval/Logos][GPLv3+]]
-- static/themes/default/icons/syncthing.png :: [[https://github.com/syncthing/syncthing/][Mozilla Public License Version 2.0]]
-- static/themes/default/icons/tahoe.png :: [[https://en.wikipedia.org/w/index.php?curid=31565848][Creative Commons Attribution 3.0 Unported License]]
