+++ data/etc/apache2/conf-available/tahoe-plinth.conf
@@ -1,11 +1,10 @@
 # Tahoe-LAFS Storage Node web interface
 
-# See if Tahoe-LAFS has basic auth
-# Enable Tahoe only from localhost
-# Do basic auth and LDAP from apache reverse proxy
-
 Listen 5678
 
 <VirtualHost *:5678>
-      ProxyPass / http://localhost:1234/
+    Include includes/freedombox-auth-ldap.conf
+    Require ldap-group cn=admin,ou=groups,dc=thisbox
+
+    ProxyPass / http://localhost:1234/
 </VirtualHost>
