+++ plinth/actions.py
@@ -98,7 +98,6 @@ import subprocess
 from plinth import cfg
 from plinth.errors import ActionError
 
-
 LOGGER = logging.getLogger(__name__)
 
 
@@ -118,14 +117,17 @@ def superuser_run(action, options=None, input=None, async=False):
     return _run(action, options, input, async, True)
 
 
-def run_as_user(action, options=None, input=None, async=False, become_user=None):
-    """
-    Run a command as a different user. If become_user is None, run as current user.
+def run_as_user(action, options=None, input=None, async=False,
+                become_user=None):
+    """Run a command as a different user.
+
+    If become_user is None, run as current user.
     """
     return _run(action, options, input, async, False, become_user)
 
 
-def _run(action, options=None, input=None, async=False, run_as_root=False, become_user=None):
+def _run(action, options=None, input=None, async=False, run_as_root=False,
+         become_user=None):
     """Safely run a specific action as a normal user or root.
 
     Actions are pulled from the actions directory.
@@ -167,7 +169,7 @@ def _run(action, options=None, input=None, async=False, run_as_root=False, becom
     if run_as_root:
         cmd = ['sudo', '-n'] + cmd
     elif become_user:
-        cmd = ['sudo', '-u', become_user] + cmd
+        cmd = ['sudo', '-n', '-u', become_user] + cmd
 
     LOGGER.info('Executing command - %s', cmd)
