--- plinth/exceptions.py
@@ -1,2 +1,0 @@
-class DomainNameNotSetupException(Exception):
-    pass
